+++ backend/controller/user.py
@@ -69,6 +69,8 @@ class BaseUser:
 
     def deleteUser(self, uid):
         dao = UserDAO()
+        if dao.getAllUserInvolvements(uid):
+            return jsonify("You erase your account because you have pending reservations."), 400
         result = dao.deleteUser(uid)
         if result:
             return jsonify("DELETED"), 200
